fix(compiler): split stack instruction analyzers
Move stack-analysis instruction behavior into focused instructionAnalyzers modules and keep analyzeInstruction as the metadata wrapper.
