+++ packages/compiler/src/stackAnalysis/instructionAnalyzers/functionEnd.ts
@@ -1,0 +1,28 @@
+import type { CompilationContext, CompilerASTLine, FunctionValueType, Stack } from '@8f4e/compiler-spec';
+import { ErrorCode } from '@8f4e/compiler-spec';
+import { getError } from '../../compilerError';
+import { stackItemMatchesFunctionValueType } from '../../utils/functionValueType';
+import { consume } from './stack';
+
+/** Consumes and validates function return values against the parsed function signature. */
+export function analyzeFunctionEnd(line: CompilerASTLine, context: CompilationContext): Stack {
+	const returnTypes = line.arguments.map(arg => ('value' in arg ? (arg.value as FunctionValueType) : undefined));
+
+	if (returnTypes.length > 8) {
+		throw getError(ErrorCode.FUNCTION_SIGNATURE_OVERFLOW, line, context);
+	}
+
+	if (context.stack.length !== returnTypes.length) {
+		throw getError(ErrorCode.STACK_MISMATCH_FUNCTION_RETURN, line, context);
+	}
+
+	for (let i = 0; i < returnTypes.length; i++) {
+		const stackItem = context.stack[context.stack.length - returnTypes.length + i];
+		const returnType = returnTypes[i];
+		if (!returnType || !stackItemMatchesFunctionValueType(stackItem, returnType)) {
+			throw getError(ErrorCode.STACK_MISMATCH_FUNCTION_RETURN, line, context);
+		}
+	}
+
+	return consume(context, returnTypes.length);
+}
