+++ packages/compiler/src/stackAnalysis/instructionAnalyzers/localSet.ts
@@ -1,0 +1,25 @@
+import type { CompilationContext, CompilerASTLine, LocalSetLine, Stack } from '@8f4e/compiler-spec';
+import { ErrorCode } from '@8f4e/compiler-spec';
+import { getError } from '../../compilerError';
+import { consume } from './stack';
+
+export function analyzeLocalSet(
+	line: CompilerASTLine,
+	context: CompilationContext
+): { consumed: Stack; produced: Stack } {
+	const consumed = consume(context, 1);
+	const operand = consumed[0];
+	const { local } = line as LocalSetLine & {
+		local: CompilationContext['locals'][string];
+	};
+
+	if (local.isInteger && operand.valueType !== 'int') {
+		throw getError(ErrorCode.ONLY_INTEGERS, line, context);
+	}
+
+	if (!local.isInteger && operand.valueType === 'int') {
+		throw getError(ErrorCode.ONLY_FLOATS, line, context);
+	}
+
+	return { consumed, produced: [] };
+}
